vcsreplicator: wait harder for topic to become available; r=sheehan
When querying topic metadata, a topic-partition may have metadata, but
it may not have a leader. This makes the topic-partition not
fully available.

This commit adds checking to ensure that broker information for
a topic's partitions is available when we wait for a topic to
come online.

This appears to reduce some race conditions in tests. But I
don't think the solution is yet a proxy for "is the topic
writable."

Differential Revision: https://phabricator.services.mozilla.com/D1323