vcsreplicator: make message aggregation deterministic; r=sheehan
Previously, we'd create a consumer bound to all partitions with
available messages and then slurp off messages until we had
exhausted a partition. There was code in _copy_messages to keep
track of which partitions were still relevant and to remove
exhausted partitions from the consumer so we didn't over-consume.

After this commit, we iterate partitions one at a time and consume
N messages from each. The new behavior is deterministic. And the
code is simpler, as we no longer have to dynamically update a
consumer with the partitions it should pull from.

As part of this change, test-sns-notifier.t gained some commands
to force messages be consumed and aggregated in a certain order. I
/think/ the test is now deterministic.

Differential Revision: https://phabricator.services.mozilla.com/D1324