vcsreplicator: use tail -n instead of cat + grep; r=sheehan
It appears there is a persistent race condition in tests between the
Kafka topic-partition becoming available and the vcsreplicator-consumer
daemon starting. This can lead to intermittent errors at the beginning
of the consumer log regarding the topic not being available.

Previously, we would `cat` the consumer daemon logs and `grep` out
unwanted lined. But this would sometimes yield extra entries in the
log related to restarting the daemon.

After this change, we only print out the final N lines of the log. We
skip over most of the early daemon log output from process startup.
Instead, we only print lines from the final, successful process start.
This should paper over any inconsistencies in daemon startup.

Differential Revision: https://phabricator.services.mozilla.com/D1325