hgserver: update check_kafka_hg_heartbeat to kafka-python 1.4; r=sheehan
It looks like kafka-python changed the API for defining the partition
leader slightly. Instead of an object being in a dict, we have an
integer. The test modified as part of this commit was failing as a
result.

We update the code to use the new API. The test now passes. But
kafka-python's error reporting is a bit more robust. So test output
changed.

Differential Revision: https://phabricator.services.mozilla.com/D1330