vcsreplicator: normalize return value from pushkey hook; r=sheehan
Mercurial 4.3 can return a boolean from its internal pushkey functions
instead of an integer. And it appears that Mercurial 4.2 is internally
inconsistent on whether a pushkey namespace returns a bool or an integer.

Consistency is good. So let's normalize the return value to an integer
on the producer side.

On the consumer side, process_hg_pushkey()'s logic currently calls
`hg debugpushkey` via hglib. Since it is executing a command, this
code always returns an integer. But the expected return value (from the
Kafka message) could be a bool. The code being executed is
`if res and res != ret`, where `ret` is our int or bool. In Python,
`True == 1` is True, so it turns out that it didn't matter whether we
recorded a bool or an int. But we want to be consistent. So let's record
integers everywhere.

Differential Revision: https://phabricator.services.mozilla.com/D1338