global: update test output for Mercurial 4.3; r=sheehan
Upgrading the hgssh and other testing environments to Mercurial 4.3
resulted in some minor changes to test output.

The printing of obsolescence activity is pretty benign.

We also uncovered a deprecation warning in the vcsreplicator consumer
code. We'll need to patch that eventually.

`hg debugpushkey` now prints bools in a few places instead of ints. We
recently worked around this on the producer side by normalizing
data written to messages to ints. But this data is output from
`hg debugpushkey` itself. We have very little control over that. So
we just accept the output changes.

Differential Revision: https://phabricator.services.mozilla.com/D1340