firefoxreleases: bump minimum version requirement; r=sheehan
We only run this extension on hgweb, which is currently running
Mercurial 4.4. We can safely drop support for old versions and
the compatibility code that came with it.

We still maintain support for Mercurial 4.3 because that's what the
local testing virtualenv is running. Hopefully that will get bumped
to 4.4 soon.

Differential Revision: https://phabricator.services.mozilla.com/D1343