vcsreplicator: use _del instead of __delitem__; r=sheehan
Our previous usage of `del bookmarks[x]` triggers a deprecation warning
in Mercurial 4.3.

In Mercurial 4.3, __delitem__ just prints a warning and calls _del. So
we change our extension to call _del directly.

There's a chance this API may go away in Mercurial 4.4. But at least it
gets rid of the deprecation warning.

Along with this change, we drop support for Mercurial 4.3. If nothing
else, this will make porting to 4.4 easier and we won't need to bother
with supporting 4.2's API.

Differential Revision: https://phabricator.services.mozilla.com/D1358