ansible/hg: version lock additional python packages (bug 1462327); r=sheehan,fubar
Currently, `yum update` is broken because we've version locked the main
Python 2.7 package. Things are complaining because python-devel and
python-libs are attempting to update but require a newer python package.

We fix this by version locking python-devel and python-libs as well.

I tested this on hgweb11 and it makes yum happy.

Differential Revision: https://phabricator.services.mozilla.com/D1354