configwizard: enable `hg wip` compatibility with hg4.6 (Bug 1461503) r=gps
hg 4.6 has renamed the "unstable" revset expression to "orphan". This
commit adds a check to configwizard that fills in the appropriate
value based on the hg version.

The reason we didn't notice this in testing is that test-wip.t only tests
that wip is enabled in the hgrc file, not that it actually works as
intended. This commit adds a basic test to ensure the revspec does indeed
return references to some commits.

Differential Revision: https://phabricator.services.mozilla.com/D1361