ansible/hg-ssh: disable bundle2 exchange of phases; r=sheehan
Mercurial 4.4 supports exchanging phases data via bundle2. Before
4.4, clients would send phases data via a "pushkey" wire protocol
command.

As part of upgrading the SSH server to 4.4, phases exchange behavior
will change and this will throw off vcsreplicator.

In order to smooth the transition to Mercurial 4.4, this change
disables the new 4.4 server behavior of allowing phases to be
exchanged via bundle2. In other words, it makes a 4.4+ server
behave like a 4.3 server.

My intent is for this hack to be temporary. Once vcsreplicator is
taught about the new bundle2 capabilities, we should be able
to remove this hack.

Differential Revision: https://phabricator.services.mozilla.com/D1376