vcsreplicator: store correct integer for boolean pushkey results; r=sheehan
The pushkey function returns True on success and False on failure.
These map to exit codes 0 and 1, respectively.

The previous code mapping bool to ints had the value inverted.

The consumer side logic only cares about the return value if the
`hg debugpushkey` invocation exits with a non-0 exit code. Since
that pretty much never happens (it only happens when there are bugs
in the replication system), our accidental inversion of the exit
code didn't matter much in practice.

Differential Revision: https://phabricator.services.mozilla.com/D1429