vcsreplicator: handle None in pushkey hook; r=sheehan
Mercurial 4.5 supports exchanging bookmarks via bundle2. It emulates
the pushkey hooks even though the "pushkey" wire protocol command isn't
being called. But as part of this implementation, it passes None as the
value for "ret" to the hook.

So, this commit teaches the replication mechanism to equate None with
0. 0 is the exit code that is supposed to mean "applied OK."

As part of this change, we introduce a test case variation for
test-pushkey-bookmarks.t that runs tests with and without using
bundle2 for bookmarks exchange. The test demonstrates (or will
demonstrate once Mercurial 4.5 is being used) that old and new clients
behave similarly.

Differential Revision: https://phabricator.services.mozilla.com/D1430