vcsreplicator: add bootstrap procedure to be run on new hgweb host r=gps
This commit adds the hgweb component of the bootstrapping procedure
on a new hgweb host. The procedure takes in JSON from the hgssh
procedure, which contains a list of repositories to replicate as
well as a set of start and end Kafka offsets. These offsets specify
a range in which the replication process can occur. We process
messages from Kafka according to the following logic: if the repo
that corresponds to the current message has already been replicated,
just process normally. If the repo has not been replicated, only
process the message if it is a repo sync message. After processing
the repo sync message we add the replicated repo to the set of
replicated repos. Once the set of replicated repos matches the
set of repos passed by the hgssh bootstrap procedure, we consider
the replication complete and we exit.

The test case extends test-bootstrap.t to take the emitted JSON
and replicate messages using this procedure. We simply pass the
JSON to the bootstrap procedure and verify the commits/hgrc files
exist on hgweb0. Since the vcsreplicator services have been turned
off on the hgweb0 host, the replication must have been the result
of the bootstrap procedure.

Differential Revision: https://phabricator.services.mozilla.com/D1156