vcsreplicator: add `--bootstrap` option to `replicatesync` r=gps
This commit adds a bootstrap option to the replicatesync command and introduces a new vcsreplicator message type, "hg-repo-sync-2". This new message type is identical to "hg-repo-sync-1" except that it includes a new "bootstrap" flag. This flag indicates if the repo sync message has been sent for the purpose of bootstrapping a new hgweb instance. If a client running the regular vcsreplicator daemon receives a message with the bootstrap flag set to true, the message is ignored. This is to avoid forcing ~4000 pulls on hosts which are already considered up to date while attempting to bootstrap. Running the replicatesync command with the bootstrap option will set this flag to True.

Differential Revision: https://phabricator.services.mozilla.com/D1383