ansible/hg: upgrade to Mercurial 4.5.3 in hgweb environments; r=sheehan
hg-web and hg-reviewboard both run a custom build of Mercurial 4.4 with
security backports from the 4.5.2 release (see bug 1440810).

This commit upgrades those environments to run a vanilla version of
Mercurial 4.5.3.

This is only part of the hgweb upgrade: we will need to update the
vendored templates so they work with 4.5. We do that in a separate
commit to make things easier to review.

Differential Revision: https://phabricator.services.mozilla.com/D1462