hgtemplates: port gitweb style changes to gitweb_mozilla; r=sheehan
This patch was mostly produced by applying the diff of the
hgtemplates/gitweb directory from the previous commit into the
gitweb_mozilla directory. There were conflicts in
changelogentry.tmpl, changeset.tmpl, map, and summary.tmpl. Of those,
only changelogentry.tmpl was non-trivial, as we've hacked up that
file a bit from upstream. But the end result was the same: basically
`<span class="logtags">` was replaced with `{alltags}`.

Differential Revision: https://phabricator.services.mozilla.com/D1466