hgmo: extension to replicate repository group owner (bug 1234556); r=sheehan
The implemented extension exposes a wire protocol command that exposes
the filesystem group owner of the repository. The extension calls this
command when performing an `hg pull` operation if it is available.

Having the group owner defined on the hgweb machines will enable us to
expose this information via hgweb in a subsequent commit.

Exposing the group owner has security implications: if you know the group
owner you know what permissions are needed to modify the repository.

But, we already expose the group owner via the "repo-group" pash command.
And, the repo group owner isn't really a secret. So I don't see any good
security justification for keeping the data hidden.

FWIW, I initially implemented this functionality via an out-of-band
mechanism that involved writing a JSON manifest of the group owners for
every repo and transferring that around. However, after thinking about
it some more, I felt that this was reinventing a wheel that the
replication mechanism already solves. Leveraging `hg pull` to keep
things synchronized feels like the right place to implement this.

Differential Revision: https://phabricator.services.mozilla.com/D1538