hgmo: add a "repoinfo" web command exposing repo group owner (bug 1234556); r=sheehan
We previously implemented support for replicating the group
owner of a repository to hgweb servers. But this data was only available
in a local file and via a Mercurial wire protocol command.

This commit establishes a new command in hgweb to expose "repository
info." It simply renders a template with custom keywords. The only
keyword we add is the repository group owner.

We implement template support in the json and gitweb_mozilla templates.
This allows the data to be requested via json-repoinfo URLs and to be
rendered in HTML, should someone want to see it in HTML.

I chose the name "repoinfo" so it is extensible. If we ever want to
expose things like the set of enabled hooks, we could do that in this
web command.

Differential Revision: https://phabricator.services.mozilla.com/D1539