robustcheckout: work around bad phases due to load balancer behavior (bug 1469351); r=sheehan
This change is a workaround for behavior in hg.mozilla.org's load
balancer triggering Mercurial to promote changesets to public when
they are canonically draft on the server. The gory details are in
bug 1462323.

See the inline comments and linked issues for more.

There is no easy way to test this. And the code is intended to be a
short-term hack. So no tests have been added.

Differential Revision: https://phabricator.services.mozilla.com/D1687