firefoxreleases: make compatible with hg46 (Bug 1469647) r=gps
Webcommand functions now only receive a single argument. So we use *args and check for the number of arguments to determine whether to expand the arguments or use the required attributes.

Webcommand functions must now also use the new response object to initiate sending a response. A "sendtemplate" convenience function is provided to send a response rendered from a template.

Differential Revision: https://phabricator.services.mozilla.com/D1711