hgmo: cat error log files in tests (Bug 1469673) r=gps
When running `hg serve ...`, there is a flag to send error output
to a file. When tests fail, they sometimes display output which
informs you there is an error, but is not useful in actually
fixing the problem (ie 500 Internal with no actual stack trace).
This commit ensures every call to `hg serve` includes an error
output file, and prints the contents of that file as output in
the test. With this patch in place any errors encountered by
the Mercurial server will be included in the test output.

Differential Revision: https://phabricator.services.mozilla.com/D1754