hgmo: enable compatibility with Mercurial 4.6 (Bug 1469673) r=gps
This commit ports the hgmo extension to be compatible with Mercurial
4.6. The main API changes are a change to the webcommand functionality.
Previously webcommand functions had the form `def wc(web, req, tmpl):`
with web being a requestcontext, req being a specific request and tmpl
being a templater object. The new API aggregates the req and tmpl
objects into the web object as web.req, and tmpl being generated
from a function.

WSGI responses are no longer automatically sent when the webcommand
functions return - instead the programmer must manually initiate a
response with the sendresponse method after specifying the body and
other fields.

The `repo[x]` syntax is being deprecated in future releases, in
anticipation of this we switch to using `scmutil.revsingle` in a few
places.

To determine if we are on a 4.6+ release we check for the `sendtemplate`
method on the requestcontext object whenever possible, and look for
`scmutil.revsingle` when this is unavailable.

Differential Revision: https://phabricator.services.mozilla.com/D1756