vcsreplicator: send a message containing all served heads (bug 1471732); r=sheehan
This commit teaches the replication system to write a new message
containing the heads in the repository after they've changed.
This message will eventually be used to determine which heads
should be exposed by the hgweb servers.

Consumers currently ignore this message completely. The presence
of this message in the replication log should be harmless. Although
for the Try repository, we will write quite a large JSON blob on
every push. The message will be similar in size to the
"hg-repo-sync" message. We know we can send and receive that message
today without issue (it's the processing it via a `hg pull`
invocation that can take minutes). So I think things will still be
safe.

Differential Revision: https://phabricator.services.mozilla.com/D1852

--HG--
extra : moz-landing-system : lando