robustcheckout: make compatible with Mercurial 4.6 r=gps
This commit makes robustcheckout Mercurial 4.6 compatible. No real code
changes were made with the exception of adding `(4, 6)` to the set
of allowed versions in the inline version check.

Most of the other changes revolved around test fixups. Almost every
test is giving devel-warn output due to core Mercurial still using a
soon-to-be-deprecated API when checking for membership via `in`. The
other change was fixing the 'server failure' test. Formerly this test
would hang in hg 4.6 due to wire protocol changes that require more than
11 bytes of transmission before being considered for a retry. I added a
conditional so the byte limit is higher when running the test on 4.6.

Differential Revision: https://phabricator.services.mozilla.com/D1888

--HG--
extra : moz-landing-system : lando