vcsreplicator: implement heads message handling (bug 1470606); r=glob, sheehan
The previous commit established a new aggregated "pending" topic
for all messages that had previously been replicated along with
a new consumer daemon for processing messages in it.

This commit teaches that new consumer daemon to handle the
"hg-heads-1" message. Whenever we receive a message listing
repository heads and the last push ID, we write a file in the
repository containing that data.

We use CBOR for the file format. CBOR is an efficient binary
serialization format. Mercurial 4.6+ has a built-in CBOR parser
and Mercurial itself is starting to use CBOR internally. So
CBOR makes sense as an interchange format. Since the CBOR module
import could make its way to random parts of the server, we add it
to all virtualenvs (even though it isn't used in all of them).

We first write out to a temporary file then do an atomic rename.
This ensures that readers see an atomic snapshot of the file content.

And because various tools (including eventually us) like to cache
files based on mtimes and because filesystem mtime often doesn't
have granularity less than 1s, we force bump the mtime if needed
to ensure the second value changes.

A test for the new functionality has been added. It foreshadows
hgweb using the heads file to determine which changesets should
be exposed. The test includes coverage for pushlog data. (We'll
also need to teach the pushlog not to expose push data for
unreplicated changesets.)

Differential Revision: https://phabricator.services.mozilla.com/D1893

--HG--
extra : rebase_source : 987567598c57984c2ca1d29f5eef56f2a73d4d5f
extra : amend_source : 2b4003a81629c17b4a3657c0b353e29803fa47e1