virtualenvs: update python-hglib in tools/bundles venvs (Bug 1470483) r=gps
This commit updates the python-hglib version to 2.6.1 in the venv_tools
and venv_bundles environments on hgssh. This brings these two environments
up to the same package version as the other virtualenvs used on hg.mo.

This change was previously deployed and then rolled back as it caused
a problem with the notification daemons. The issue comes from the
way `hglib` converts arguments in the `cmdbuilder` function to
command line arguments. In version 1.9, the kwargs are converted to
option arguments first, and the positional arguments are appended to
the command after. So

`hglib.util.cmdbuilder(b'log', b'--hidden', r=revs, template=template)`

produces

`log -r <revs> --template <template> --hidden`.

However in some version between (or including) 1.9 and 2.6.1, an extra
guard was added to ensure that positional arguments are explicitly
not treated as flags. So the same call in v2.6.1 produces instead:

`log -r<revs> --template <template> -- --hidden`.

Note the extra `--`. This argument causes hg to consider every argument
that follows as a positional argument. Thus, the error saying "--hidden
is not under root" is hg telling us that it can't show the history of
the file `--hidden`, so it must not be under the root of the directory.

Simply changing the positional argument to a keyword argument `hidden=True`
will cause the argument to be treated as a flag and added to the command
before the "--".

Differential Revision: https://phabricator.services.mozilla.com/D1952

--HG--
extra : moz-landing-system : lando