vcsreplicator: log exceptions instead of raising r=gps
While running the bootstrap procedure on hgweb1/2 in MDC1, we noticed
logs stop emitting from the process after a certain amount of time.
We believe the current implementation of raising/hard failing on the
main thread when something goes wrong is the cause of this issue. It is
possible that the main thread raises an uncaught exception and exits,
yet the ThreadPoolExecutor is still alive and performing the clones. However
since the main thread is dead, the logger is also dead and therefore
no more logs are being produced.

Differential Revision: https://phabricator.services.mozilla.com/D1974

--HG--
extra : moz-landing-system : lando