vcsreplicator: only expose replicated revisions via hgweb (bug 1470606); r=glob, sheehan
Now that we write out a file containing the set of fully replicated
data in a repository, it's now time to put it to use.

This commit implements a new Mercurial extension that provides a new
repoview/filter that takes the replicated-data file into account.

A repoview is simply a function that returns a set of revisions that
should be hidden. By default, hgweb uses the "served" repoview. This
view filters out hidden/obsolete and secret revisions.

We implement a "replicatedserved" repoview. It starts by collecting
the revisions filtered by the "served" repoview. If the replicated
data file exists and can be parsed, we perform a DAG traversal
starting at all DAG heads and stopping at revisions specified in
this file and revisions hidden by the "served" repoview. These
revisions are unioned with the revisions from the "served" repoview to
form the new set of hidden changesets.

The parsed data file is exposed as a property on the repo object. This
will allow other extensions (such as pushlog) to access it.

Test changes demonstrate that unreplicated changesets are no longer
exposed via hgweb! However, the pushlog assumes that !visible=obsolete
and it marks the unreplicated changeset as obsolete. This is obviously
not desired. A subsequent commit will teach the pushlog code to do
something reasonable.

Because the pushlog demonstrates buggy behavior, deploying this without
pushlog fixes should probably not be done. If this were to be deployed,
a client consuming the pushlog could see a changeset marked as obsolete
then later as non-obsolete. And pushes would go from having an empty
list of changesets to a populated list. This would almost certainly
confuse pushlog consumers.

Differential Revision: https://phabricator.services.mozilla.com/D1894

--HG--
extra : rebase_source : 8922faeb8d51a9da9612d1574b14dd8a034383a3
extra : amend_source : c2b4f4351dfdb71ffee0e06ad6a158b5c16853be