vcsreplicator: add command to delete repositories from all mirrors (Bug 1464166) r=gps
This commit adds a command to the vcsreplicator extension called
`replicatedelete`. This command will delete the repository on which the
command is run, and then publish a Kafka message to instruct the hgweb
mirrors to remove the repository as well.

On hgssh, the repo is renamed to "<repo>.todelete", making the repo
unwritable, and then the newly named repo is deleted. On hgweb, the repos
are moved to "/repo/hg/todelete/repo" and are then deleted.

Differential Revision: https://phabricator.services.mozilla.com/D1451

--HG--
extra : amend_source : c99f397f424b9ca5d98e224518e990cd28708985