vcsreplicator: collect Kafka messages one partition at a time; r=gps
As discussed in review, collecting the set of bootstrap messages
from a single Kafka consumer subscribed to all topics create unnecessary
complexity in the code. It is simpler to assign a consumer to each
partition one at a time and pull the messages for that partition until
we reach the limit of the bootstrap range. This commit changes
the message collection phase of the target bootstrap to do just that.

Differential Revision: https://phabricator.services.mozilla.com/D1954

--HG--
extra : amend_source : 7e5738a9271410094e80eae239fba3fff1f4658c