vcsreplicator: emit JSON with auditing information from hgweb bootstrap r=gps
This commit adds output to the hgweb bootstrap procedure that contains
repositories that errored out in the process of cloning or applying
extra messages in the bootstrap range. The output is a JSON document
mapping repo names to an array of strings with details of the failure.

Differential Revision: https://phabricator.services.mozilla.com/D2005

--HG--
extra : rebase_source : badc333e6f41b12b2f2f00c28beca4493d61892d