pushlog: pass database handle into query function; r=sheehan
Previously, the PushlogQuery instance held a handle on the database
connection. The connection was only used by DoQuery().

Passing the connection into DoQuery() will help prevent a leak
of the connection via PushlogQuery instances and will help
us limit the lifetime of the connection to only when it is needed.

Differential Revision: https://phabricator.services.mozilla.com/D1979

--HG--
extra : moz-landing-system : lando