pushlog: use existing API for accessing pushlog connection; r=sheehan
The repo.pushlog object was invented after this code was written.

repo.pushlog.conn(readonly=True) works exactly like what the
code deleted by this commit was doing.

This patch starts the process of getting pushlog-legacy to use
the new API provided by the pushlog extension. The end goal is
for most of the business logic to be in the pushlog extension
itself.

Differential Revision: https://phabricator.services.mozilla.com/D1982

--HG--
extra : moz-landing-system : lando