vcsreplicator: add last push ID to heads message (bug 1470606); r=sheehan
Our intent is to use the heads message as a coordination event to
control what data is exposed on hgweb machines. As part of testing
this eventual mechanism, I discovered that the pushlog needs to
be integrated with this new mechanism because it assumes hidden
changesets are obsolete.

This commit adds the last push ID from the pushlog to the heads
replication message, if available.

Strictly speaking, this change is backwards incompatible and we
should introduce a new message to convey the new data. However,
nobody is using the heads message yet. So we can get away with this
change without introducing a new message type.

Differential Revision: https://phabricator.services.mozilla.com/D1931

--HG--
extra : moz-landing-system : lando