pushlog: use subquery when obtaining pushes; r=glob
Upcoming commits will teach the pushes() method to filter what data
is returned. Because we need to eventually use LIMIT and OFFSET
and because we're using a LEFT JOIN, we need to use an inner SELECT.
This is because LIMIT and OFFSET apply at the SELECT level and if
we applied them after the LEFT JOIN, we'd be operating against node
counts instead of push counts.

So this commit refactors the query to use an inner SELECT. It looks
like bad SQL now. But it will make sense once we're doing more
filtering in future commits.

Differential Revision: https://phabricator.services.mozilla.com/D2013

--HG--
extra : moz-landing-system : lando