ansible: add `--no-setuptools` to mozreview-virtualenv get-pip invocation r=gps
`get-pip-8.0.3.py` includes some logic to determine if it is being asked
to install several packages implicitly by the user (pip, setuptools and
wheel). The script will try and import the package, and if it succeeds
it assumes the user does not want the package installed unless it finds
the package on the command line later in the script. Previously, the
script would find setuptools, set the value to False and later on install
the version we specify on the command line. However it seems the default
downloaded version of setuptools now performs a check of `sys.version_info.major`,
which was added in Python 2.7, so when we run the script under 2.6 an
Exception is raised which causes the image to fail to build.

Adding the `--no-setuptools` command line argument will let us bypass
the import check by telling the script we do not want setuptools to
be implicitly installed. Shortly after the previously failing line
of code, get-pip runs `pip install` with setuptools explicitly set to
version 20.1.1.

Differential Revision: https://phabricator.services.mozilla.com/D2054

--HG--
extra : moz-landing-system : lando