ansible: playbook to perform a stream clone to hgweb mirrors
These are the commands I've been using to try to get repos to
reproduce with exact `hg mozrepohash` on hgweb machines. The
current playbook nukes the repo from all hgweb servers. You'll
probably want to hack up ansible/hosts or the playbook to not
do that on servers that are in the load balancer when operating
on important repos.