hghooks: enable firefoxreleases extension in tests; r=sheehan
The firefoxreleases extension defines the mozilla.firefox_releasing
config option. But it defines it as dynamicdefault, which means we
need to supply a default value to ui.configbool() or else None is
returned. So we change consumers of this config option to pass a
default variable.

The usage of this config option is now consistent throughout the repo.
Although it should arguably not be using dynamicdefault. But we
can deal with that another time.

Differential Revision: https://phabricator.services.mozilla.com/D2064

--HG--
extra : moz-landing-system : lando