vcsreplicator: invalidate repositories when .hg/replicated-data changes (bug 1475656); r=glob
@localrepo.repofilecache doesn't do what its docstring says it does.
The docstring says that the function result will be cached for as long
as stat() on the listed file(s) returns the same result. In reality,
the function result is cached until the cache is cleared, at which
point a stat() will be performed on next access and the cached result
will be returned, if possible. And the cache is only cleared when
the localrepository object is mutated by itself.

(I have submitted a patch upstream to have the docstring reflect
reality.)

In the context of hgweb, localrepository instances are cached between
requests. And there's a list of files in mercurial.hg.foi that
are stat()ed to determine whether a cached localrepository instance
can be used for a subsequent request.

Without .hg/replicated-data in mercurial.hg.foi, this file could change
on disk and a subsequent HTTP request could use the cached value of the
replicated_data attribute. This would lead to badness. The old cached
value would linger until either:

a) the cached repo was evicted from the repo cache
b) the hgweb process exits
c) any other path in mercurial.hg.foi changes (they would change when
   new changesets are added to the repo)

Differential Revision: https://phabricator.services.mozilla.com/D2147

--HG--
extra : moz-landing-system : lando