vcsreplicator: only log "successfully cloned" if clone is actually successful r=gps
Previously the design of the bootstrap process called for a
"hard fail" in the event a clone was unsuccessful. We have since
changed this to instead continue with the bootstrap process but
have errors logged and recorded for future auditing. The log
message saying "successful clone" should be conditional on whether
or not an error is added to the audit output.

Differential Revision: https://phabricator.services.mozilla.com/D2189

--HG--
extra : moz-landing-system : lando