Bug 1446453 - serverlog: port to Mercurial 4.6; r=sheehan
The wire protocol code changed substantially in 4.6. This required
a number of conditional checks throughout the extension.

It's worth noting that we now record the response size for all HTTP
requests. This is because prior to 4.6, some HTTP handlers were using
the legacy request.write() API to send response bytes. This extension
isn't monitoring this API. Post 4.6, all I/O is performed via a
generator, which is monitored.

Differential Revision: https://phabricator.services.mozilla.com/D2207

--HG--
extra : moz-landing-system : lando