qimportbz: fix test under Mercurial 4.6.2 (Bug 1477519) r=gps
This test has been broken for some time. It turns out bzexport changes
the patch name under the hood and this was not reflected in the test.
As a result, the default patch name already existed and the test
would enter an endless loop asking to rename the patch.

Differential Revision: https://phabricator.services.mozilla.com/D2298

--HG--
extra : moz-landing-system : lando