pushlog: append '/' to `url` pushlog template keyword in 4.6+ (Bug 1470489) r=gps
It turns out the conditional test output added in previous commits
for this bug was being generated due to the absence of a WSGI
config in the `hg serve` instances. This config usually sets a
`paths` config section, where local filesystem paths can be
mapped to a web url under which the repo should be served. For
example on hg.mo, we would map "/repo/hg/mozilla/" to "/",
so that "/repo/hg/mozilla/mozilla-unified" would appear as
"https://hg.mozilla.org/mozilla-unified" as a web facing
url. This commit adds a `wsgiconfig` function to generate
a config file which can be passed to the `hg serve` instance
via the `--web-conf` option.

As a result of this change, all urls referencing the serve
instance without the local repo name have had the repo name
appended (ie "https://localhost:$HGPORT" is now
"https://localhost:$HGPORT/reponame").

After making this change, I noticed that the base URL on 4.6.2
no longer ended with a '/', which caused links to revisions and
other resources to 404. This commit also fixes this by appending
'/' to the variable that is later seeded into the `url` template
keyword when running on 4.6+.

Differential Revision: https://phabricator.services.mozilla.com/D2270

--HG--
extra : moz-landing-system : lando