qimportbz: Make compatible with modern Mercurial versions (Bug 1477519) r=gps
Under Mercurial 4.6.2, `patch.extract` is now a context manager.
Callers no longer need to unlink or remove the file, as it is
tied to the lifetime of the context manager. This commit makes
qimportbz use the context manager when running Mercurial 4.6.2.

We also register config options, to avoid devel-warnings in tests.

Differential Revision: https://phabricator.services.mozilla.com/D2299

--HG--
extra : moz-landing-system : lando