hghooks: fix releasing repo configuration in hooks tests r=gps
Several of Mozilla's Mercurial hooks are designed to run on a
subset of repositories that meet a specific criteria. The main
example of one of these predicates is the "Firefox releasing
repo", which is a repo from which a release version of Firefox
is built. For some time, hooks tests which run on these repos
have been failing. After some digging I realized that the flags
and config settings we use to identify these repos were incorrectly
configured in the test suite.

This commit adds a `configurereleasinghooks` function, which creates
an hgrc config and `IS_FIREFOX_REPO` file that specifies the repo
is a releasing repo. After this change, all tests but one are now
passing.

Differential Revision: https://phabricator.services.mozilla.com/D2323

--HG--
extra : moz-landing-system : lando