ansible/hg-ssh: use pip-compile for bundle requirements; r=sheehan
pip-compile can be used to produce pip requirements for us. Let's
bring requirements-bundles.txt under the management of pip-compile.

We don't actually need the --find-links entries in the requirements
file. If we introduce a custom Mercurial build, we will. But that is
a special case and we can ignore that for the time being IMO.

All pinned versions were carried over for backwards compatibility.

Differential Revision: https://phabricator.services.mozilla.com/D2338

--HG--
extra : moz-landing-system : lando