hgtemplates: add test for templates differences; r=sheehan
Currently, templates changes from one Mercurial version to the
next are applied manually. This process is brittle and prone to
mistakes.

We want to make this process more resilient and take less time.
Our solution to that will be to track our changes to the canonical
templates in patch files and to use a script to automate the process.

The test added in this commit essentially captures the process we
use for upgrading templates. It demonstrates that there are
a handful of one-off actions (which could be scripted later) and
that the final output diverges from what's checked into v-c-t.

Subsequent commits will improve the situation.

Differential Revision: https://phabricator.services.mozilla.com/D2377

--HG--
extra : moz-landing-system : lando