vcsreplicator: make replicated_data a regular @property (bug 1475656)
I have a hunch that @localrepo.repofilecache may be conspiring
against us. Let's use a @property to isolate it.

This will result in a minor performance loss due to having to
read the .hg/replicated-data file on every property access.
But that file is at most ~50KB (on Try - and growing) and the
property is only accessed when computing the repoview and
calculating the last replicated push ID. So I'm optimistic
the performance loss will be minimal.

--HG--
extra : histedit_source : e771de0d1a6cd8ce9326d3e167a8008e56612de4