hgtemplates: add script for managing templates; r=sheehan
test-gitweb-template-patches.t captures a lot of the actions we
perform to upgrade templates. We want to automate that process.

This commit begins the process of automating template updates by
introducing a script to do that.

The script is currently pretty simple: it merely copies the
canonical templates and removes unwanted files from the new
templates directory.

The lack of test changes demonstrate feature parity.

Future commits will add more functionality to the script.

Differential Revision: https://phabricator.services.mozilla.com/D2378

--HG--
extra : moz-landing-system : lando